scraper/requirements.txt
pandas==2.2.3
openpyxl==3.1.5
python-dateutil==2.9.0.post0
selenium==4.27.1
